move more stuff, fix GHA workflows
