+++ .github/workflows/test_sbml_semantic_test_suite.yml
@@ -8,9 +8,9 @@ on:
   pull_request:
     paths:
       - .github/workflows/test_sbml_semantic_test_suite.yml
-      - python/amici/ode_export.py
-      - python/amici/sbml_import.py
-      - python/amici/import_utils.py
+      - python/sdist/amici/ode_export.py
+      - python/sdist/amici/sbml_import.py
+      - python/sdist/amici/import_utils.py
       - scripts/run-SBMLTestsuite.sh
       - tests/testSBMLSuite.py
       - tests/conftest.py
