--- python/amici/MANIFEST.template.in
@@ -1,1 +1,0 @@
-include *.cpp *.h
